pyproject.toml
[project]
name = "neural-operators"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "neuraloperator>=2.0.0",
    "torch>=2.10.0",
]
